requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via django
attrs==23.1.0
    # via jsonschema
black==23.3.0
    # via -r requirements.in
build==0.10.0
    # via pip-tools
click==8.1.3
    # via
    #   black
    #   pip-tools
coverage[toml]==7.2.3
    # via pytest-cov
django==4.2.10
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-modeltranslation
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.2.0
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-filter==23.5
    # via -r requirements.in
django-modeltranslation==0.18.9
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.26.2
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.1.1
    # via pytest
flake8==6.0.0
    # via -r requirements.in
freezegun==1.4.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements.in
jsonschema==4.17.3
    # via drf-spectacular
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
numpy==1.24.2
    # via pandas
openpyxl==3.1.2
    # via -r requirements.in
packaging==23.1
    # via
    #   black
    #   build
    #   pytest
pandas==2.0.0
    # via -r requirements.in
pathspec==0.11.1
    # via black
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.2.0
    # via black
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.6
    # via -r requirements.in
pycodestyle==2.10.0
    # via flake8
pycryptodome==3.20.0
    # via -r requirements.in
pyflakes==3.0.1
    # via flake8
pymemcache==4.0.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.19.3
    # via jsonschema
pytest==7.3.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   freezegun
    #   pandas
pytz==2023.3
    # via
    #   djangorestframework
    #   pandas
pyyaml==6.0
    # via drf-spectacular
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   pytest
typing-extensions==4.5.0
    # via django-modeltranslation
tzdata==2023.3
    # via pandas
uritemplate==4.1.1
    # via drf-spectacular
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
